first test cases implemented (and already fruitful - found a few bugs); a new restraint/constraint/tables atom list management is proposed - the benefits are that the lists can be expanded in place and keep the original data for macro-lists like C1_tol, C1_+ etc.. will plug it in later...

+++ trunk/sdl/talist.h
@@ -163,9 +163,9 @@ public:
   }
 //..............................................................................
   // cyclic shift to the left
-  void ShiftL(int cnt)  {
+  void ShiftL(size_t cnt)  {
     if( FCount == 0 )  return;
-    int sv = cnt%FCount;
+    size_t sv = cnt%FCount;
     if( sv <= 0 )  return;
 
     if( sv == 1 )  {  // special case
@@ -187,9 +187,9 @@ public:
   }
 //..............................................................................
   // cyclic shift to the right
-  void ShiftR(int cnt)  {
+  void ShiftR(size_t cnt)  {
     if( FCount == 0 )  return;
-    int sv = cnt%FCount;
+    size_t sv = cnt%FCount;
     if( sv <= 0 )  return;
 
     if( sv == 1 )  {  // special case
